implemented preperations for changing the simulation config during runtime; ms per time step can be changed during runtime

+++ microtrafficsim-core/src/main/java/microtrafficsim/core/logic/Node.java
@@ -144,7 +144,7 @@ public class Node implements IDijkstrableNode {
 			Iterator<AbstractVehicle> iter = assessedVehicles.iterator();
 
 			// get vehicles with max prio
-			byte maxPrio = Byte.MIN_VALUE;
+			int maxPrio = Integer.MIN_VALUE;
 			while (iter.hasNext()) {
 				AbstractVehicle v = iter.next();
 				// if
